% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lq.R
\name{lq_tag_rm}
\alias{lq_tag_rm}
\title{Remove a tag from a template}
\usage{
lq_tag_rm(path, name, tag, source = NULL)
}
\arguments{
\item{path}{Path to the library.}

\item{name}{Template name.}

\item{tag}{Tag name.}

\item{source}{Which source to target. Required if the library has more
than one source.}
}
\value{
Invisibly, \code{NULL}.
}
\description{
Remove a tag from a template
}
\examples{
store <- withr::local_tempdir()
lq_init(store)
lq_add(store, "letter.md", "v1")
lq_tag(store, "letter.md", "final")
lq_tag_rm(store, "letter.md", "final")
}
